+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This file is used to list changes made in each version of grafana.
 ## Unreleased
 
 - Added support for HTTPS endpoint
+- Added support for endpoints with url_path_prefix
 
 ## 9.2.1